core/api-spec: use "is_local" instead of "origin"
This commit proposes to replace the "origin" field with "is_local".
The value of this field is either the string "yes" (equivalent to
the previous "local") or "no" (equivalent to the previous
"external").

Notes:

- "external" feels weird in some contexts, especially if we carry
this concept forward to the transaction level.
- we can't use booleans due to restrictions in the indexer backend.
- we use "is_local" instead of "local" because it flows better in
the context of a tx input or output, where it is prefixed with
"asset_", i.e., "asset_is_local".

Closes chain/chainprv#1901