core/api-spec: add "is_local" field to transaction & unspent output
When connected to a network with other cores, we will want to differentiate
transactions involving the core. To achieve this with filters, we will add
`"is_local": <"yes"|"no">` to the transaction object (transaction level,
inputs, and outputs) and unspent output object.

Each object `is_local` if:

1. INPUT: `action` is `"issue"` AND `asset_is_local` is `"yes"`, OR
`action` is `"spend"` AND `account_id` is not null.
2. OUTPUT: `action` is `control` AND `account_id` is not null.
3. TRANSACTION: for any input or output, `is_local` is `"yes"`

Outputs where `action` is retire will not have the `is_local` field.

Closes chain/chainprv#1884