wercker: consolidate all steps
Because wercker doesn't run pipelines in parallel, and because it has to do
all of its setup and teardown and archiving and whatever else for each
pipeline, it's actually much slower to have all of our steps separately.
This is a proposal to consolidate them. 

(Perhaps Dashboard should continue to be its own status?) 

/cc @erykwalder @kr 

Closes chain/chainprv#1923