core: use sensible default for /list-balances sum_by parameter
This commit applies a sensible default of ["asset_alias", "asset_id"]
for the /list-balances sum_by parameter. This is likely to be the
desired outcome a large majority of the time, and having it as a default
makes educating users about /list-balances much easier. The sum_by
parameter goes from being a source of confusion to a parameter only
used for esoteric use cases.

Side effects:

- Since there is default behavior, it is no longer meaningful to provide
an 'addSumByParameter' method in the Java SDK, since any client input
will override the default.
- The dashboard no longer needs to provide these values as a default.
The placeholder text is left as-is as an education device.

Resolves chain/chainprv#1777 

Closes chain/chainprv#1914