protocol/patricia: slice keys judiciously
Interior nodes of the patricia tree re-slice the keys of the leaves
inserted into the tree. The keys are relatively large (~340 bytes).
Previously, if we deleted a leaf, the keys of the interior nodes
would still reference the underlying array of the leaf's key,
keeping the entire old key in memory.

With this commit, all nodes on the path to the inserted key use
the new node's key. All nodes on the path to a deleted key use
the other child's key.

I tried testing this locally by inserting 20,000 values, deleting a
random 10,000 values, forcing a GC and logging `runtime.MemStats`.
This version used about 10% less memory after the GC. If you delete
older keys first (instead of random), the savings is closer to 20%.

Closes chain/chainprv#1929