core/api-spec: add network version to /info response
This commit proposes the following:

- A hardcoded global constant indication the network RPC version of
the core. Cores with identical values for the network RPC version can
connect to each other via network RPCs. Otherwise, they are
incompatible. Breaking changes in the syntax or semantics of the
network RPCs should result in an increment of the network RPC version.
This is not directly related to any overall product versioning
system we may adopt in the future. For builds up to 10/24, the
network version should be 1.
- Exposing the network version via the /info client RPC. This
allows cores to determine whether they are compatible. In
particular, this will be used to determine whether a locally-
installed Chain core can connect to Testnet.

Closes chain/chainprv#1887