Code of Conduct
I'd like to propose a code of conduct for Chain's open source community. I
know a lot of folks at Chain care about inclusion and kindness and I think
the Contributor Covenant project blends well with our style and values.

We'll establish a limited, private email group for folks of the community
to report violations or general issues related to the ideas outlined in the
guide.

I also liked Go's anonymous contact form for reporting issues:
https://docs.google.com/a/32k.io/forms/d/e/1FAIpQLSczOn135aILRITeQqH-YaurBL69g_pO-Ham9_GwKlFVgnyZZw/viewform

Closes chain/chainprv#1927