cmd/md2html: interpolate code
This path adds the ability to interpolate code in markdown files.

```
$code absolute-file-path snippet-name
```

When parsing markdown files, if md2html encounters a `$code` identifier, it
will read the code snippet (the entire file if snippet-name is committed)
and insert it into the markdown file.

Snippets can be defined in a source file use `snippet name` and
`endsnippet` keywords.

```
//snippet main
func main() {}
//endsnippet
```

See md2html code documentation for more details.

Closes chain/chainprv#1947