cmd/testbot: run tests on pull requests and report as GH status
`testbot` has been updated to run integration tests when a PR is opened
or synchronized (pushed to). The `benchcore` command will still be
triggered
by commits to `main`. `testbot` will report results of tests to GH as a
status.
Upon failure, the test output will be uploaded to s3 and linked in the
status.
For now, the tests will not be required for merge.

Future work involves creating a new webhook that is triggered by pull
requests
and updating the current webhook's endpoint.

Closes chain/chainprv#1837