core/api-spec: replace build-tx's final param with allow_additional_actions
This commit proposes two updates to /build-transaction:

1. The "final" parameter (currently undocumented) should be replaced
with "allow_additional_actions", which has the inverse semantics of
"final". When this parameter is set to true, the generated signature
programs will commit only to inputs, outputs, and transaction metadata
present in this build call. Requests that omit this parameter or set it
to false will generate signature programs that commit to the full tx
hash. In this case, build-tx should return an error if the inputs and
outputs do not balance.

2. Setting transaction reference data should be performed via an action.
An error should be returned if this action is present and the
transaction reference data is alread set.

Closes chain/chainprv#1967