core/signers: use slices of bytes for path
chainkd's Child function takes a byte slice to derive the child, and is
not limited to the size of a uint32. As a result, there is no need for a
composite key index at the level of accounts, assets, or control
programs. Additionally, the namespace and account/asset part of the
derivation path can be collapsed into one byte slice.

CPU profiles of benchcore have shown that a non-significant amount of
time is spent on EC math, including deriving keys. This change
reduces the number of derivations for a control program key from 5 to 2,
and for an asset or account from 3 to 1.

Closes chain/chainprv#1996