crypto/ed25519/chainkd: don't sort derived keys
Sorting derived keys shuffles the placement order of pubkeys in an
issuance or control program. This makes the order of the signatures
unpredictable. By avoiding this shuffling, txbuilder.Materialize does
not need to determine the position of signatures in an input, but can
just trust that they match the order of xpubs.

Closes chain/chainprv#2029