protocol: make validation.ConfirmTx read-only
`validation.ConfirmTx` used to be side-effect free but issuance memory
pruning snuck in. This commit moves pruning to where other modifications of
the snapshot happen to keep `ConfirmTx` a pure function.

This should also be marginally faster because we don't need to iterate over
all of the issuance memory for every tx in the block.

Closes chain/chainprv#2069