Revert "core: use mempool"
This reverts commit 384e7f53ca8dc799bb28bae3ea27707b1fbe0308.

We can't use a mempool quite yet because all processes within a Core
need to be able to accept a submitted tx. We can switch to a mempool
once we have support for gossiping between processes within a Core.

Instead, we can improve the performance of txdb by relaxing the
consistency requirements and marking the `pool_txs` table as
`UNLOGGED`.

https://www.postgresql.org/docs/9.1/static/sql-createtable.html#AEN67445

Closes chain/chainprv#2105