core: refactor ServeMux
This reorganizes the code so that all RPC handlers are
registered in the same http.ServeMux table. Future
code will be able to use this table to determine if a
request corresponds to a legitimate request path before
it actually needs to call the handler. For example,
metrics code can create latency histograms on demand
but it shouldn't create those objects for bogus
requests. It can use the ServeMux as a lookup table to
determine which request paths are okay.

Closes chain/chainprv#2127