core/txdb: don't lose txs during Store.Dump
Instead of selecting all of the pool txs and then truncating the
entire table, delete only the rows we selected. This prevents the
race condition where a tx is inserted between the SELECT and the
dump.

It's okay to drop pool txs because they'll be re-submitted by the
submit handler, but it'll increase latency to confirmation for the
dropped txs.

Closes chain/chainprv#2118