cmd/benchcore: bump up to m3.xlarge
Running on a full day's worth of UTXOs, the m3.large exhausts its
storage on the postgres instance. This bumps up to the m3.xlarge and
uses the mounted drive for postgres.

Also, tee the output of cored to a local file for debugging and provide
the GOTRACEBACK=crash environment variable to generate a core dump on
crash.

Closes chain/chainprv#2137