core: always allow local connections
This is a simpler mechanism that also provides a better
user experience.

This rests on the assumption that Chain Core Developer
Edition is always deployed on a single-purpose system:
either a laptop, or a dedicated VM, or a dedicated
container (with its own network namespace); not on a
multi-user system.

Further work: corectl command to generate auth tokens.

Closes chain/chainprv#2132