protocol/validation: fix 32-bit build
This comparison won't compile when int is 32 bits.
Luckily, the protocol says the max value for this list
is actually the max *signed* 32-bit int, and the
existing check was looser than necessary. We can fix two
birds with one stone.

Note that this boundary check is redundant with a
similar check in the deserialization logic.

Fixes chain/chainprv#2169.

Closes chain/chainprv#2172