docker/de: generate access token during startup
Docker for Mac's networking configuration doesn't allow
connections to originate through the host machine's localhost,
forcing the need for a client access token. This commit generates
a client access token during startup. The token is printed to
stdout as well as logged to a file.

Also, the startup logs have been modified to address chain/chainprv#2189 

Closes chain/chainprv#2177