cmd/cored: support configure and reset on windows
We use the exec syscall on Unix to reset in-memory data whenever we reset
or configure the database. But exec is unavailable on Windows, so we need a
different approach there. The cored process started by the user starts a
child process to do the actual Chain Core functions, and the parent merely
monitors the child and restarts it as necessary. The child then exits
instead of calling exec on itself.

Closes chain/chainprv#2154