sdk/java: persist basic authentication for mock hsm urls
The MockHsm.Key object holds a mock hsm url. When
generating these urls, the current context's url is appended to.
This change assures that if an access token has been set on the
context, the resulting mock hsm url includes the access token.

The Context class has been modified to allow basic authentication
within the url.



Closes chain/chainprv#2239