core: refactor action construction
Refactor the build endpoint so that actions are constructed in their
respective packages. This is working towards the goal of:

* removing package globals (ex, in `core/asset`, `core/account`)
* removing dependency on the database on the Context

In that world, we won't be able to just unmarshal an action and have it
be capable of building itself. It'll need a reference to its package's
storage/database.

This is also a little cleaner in terms of package boundaries because the
`core` package only knows that these actions exist. The details and the
structs are not exported from their respective packages.

Closes chain/chainprv#2228