cmd/md2html: deal with working dir relative path issues
The current md2html workflow is to 1) serve docs by running md2html from
$CHAIN, and 2) compile docs by running md2html from doc/. This means
that the rendering behavior must remain consistent even when the working
directory changes.

The tool has been updated in two locations:

1. It no longer assumes layout.html is in the working directory, and
instead performs a search for a layout file, starting with the
pre-rendered file's parent directory, and ending with the working
directory.
2. Code snippet source paths are now assumed to be relative to the
pre-rendered file into which the snippet is being interpolated, rather
than relative to the working directory. Absolute paths are assumed to
be "relative" to $CHAIN.

Closes chain/chainprv#2247