core: omit stack trace on non-500 errors
It looks like this might've been the intended behavior since a stack trace
is explicitly added for 500s. Most of our errors are wrapped with a stack
trace. so `log.Write` would always print that.

If we don't like this, we can also go the other direction with a
blacklist of errors that we don't want stack traces for.

Closes chain/chainprv#2282