net/http/authn: don't return WWW-Authenticate header
TL;DR: this helps to fix the dashboard in MS Edge.

The WWW-Authenticate header is used to instruct user agents (usually web
browsers) to present a generic username/password challenge to the human
user. So far, this has not been an issue, but it appears that recent
builds of Microsoft Edge show this challenge for _all_ outgoing
requests, even those that originating from JavaScript. This behavior is
equivalent to a bug for the dashboard, as the dashboard expects to handle
401 responses with custom behavior, rather than a Microsoft-designed
password challenge.

The workaround offered doesn't really have drawbacks. We don't typically
expect users to interact directly with the client API via a standard
browser window. In-browser apps are unaffected by this change; in
MS Edge this change actually _fixes_ buggy behavior.

As a final note, eliminating username/password challenges is probably a
good thing from a usability standpoint, since our access token format
(id:secret) somewhat intentionally obfuscates the distinction
between usernames and passwords; Chain Core users should never have to
think about credentials that involve two different values.

Closes chain/chainprv#2295