core/asset: encapsulate storage
Store a db on the `asset.Registry` struct and use that instead of
pulling a db handle out of the context. There are a few places where we
need to pass our db handle to another package via the context
(pg.ForQueryRows and core/signers functions), but those can be fixed up
when those packages are fixed up.

Closes chain/chainprv#2326