protocol: run block callbacks in parallel
We can't run block callbacks asynchronously yet because it changes
the semantics about what it means for a block to land (ex, the account
UTXOs might not be indexed yet). However, we can at least run them in
parallel.

This requires all tests that commit blocks on the Chain to use 
pgtest.NewDB instead of pgtest.NewTx. I think I got them all, but 
there could be a few more.

According to the benchcore runs I just ran, this gets us from ~200 tx/sec
to ~300 tx/sec. It should have a similar impact when catching up to a
a generator that's far ahead.

```json
{
	"commit": "9ade5bf0aa4618bda9b90e7fef96cd00e442c509",
	"elapsed_ms": 709931,
	"finished": "2016-10-14T20:04:34.77882718Z",
	"prog": "IouSettlement.java",
	"txs": 216000
}
```

Closes chain/chainprv#2350