sdk/java: refactor batch methods
This commit introduces the BatchResponse class to the Java SDK,
which formalizes batch response handling. The BatchResponse class
separates response objects in a batch response into successes and
errors, providing the user with an easy interface to deal with
any errors that may be interleaved with successful results from a
batch API call.

This reimplements singletonBatchRequest in terms of a new method,
batchRequest.

This commit builds off of major work from @boymanjor. It
supersedes chain/chainprv#2367 and chain/chainprv#2126.

Closes chain/chainprv#2367
Closes chain/chainprv#2126

Closes chain/chainprv#2381