core: add authentication to /debug/ endpoints
Don't use the http.DefaultServeMux in cmd/cored. In
`(*core.Handler).init`, setup explicit handlers for the
`net/http/pprof` endpoints and the `expvar` endpoint so that they'll
be wrapped in client-token authentication.

For now, we must copy the `expvar` handler because it's not exported. It
will be exported in Go 1.8:
https://go-review.googlesource.com/#/c/24722/

Resolve chain/chainprv#2198.

Closes [unavailable]