cmd/cored: include processID in /debug/vars
It'll be helpful when funneling metrics to librato to be able to use
the cored's processID as the source of metrics.

Originally, I was going to use the URL of cored, but in a multi-process
cored you might want the metrics daemon to sit alongside cored on the
same box, accessing it through localhost. Then all the processes would
appear as the same source.

Closes [unavailable]