sdk/java: update HsmSigner and MockHsm intefaces
This commit updates the HsmSigner and MockHsm 
classes to utilize contexts instead of hsm URLs. The
HsmSigner now keeps a map of contexts to xpubs.
The MockHsm.Key no longer has an hsmUrl member.
It now has a method to generate a context that can
make requests to its mock hsm.

Closes [unavailable]