sdk/java: use ArrayList constructor when updating lists
This commit changes all set methods that update lists
to use ArrayList<>(List) when setting the attribute from
an existing list. This is motivated by a bug I found when
passing the result of Arrays.asList to QueryBuilder.addAlias.
Arrays.asList returns a fixed length array that cannot be modified.

Closes [unavailable]