docs: replace index.md with index.html
We'd like to serve a custom index page for the root of /docs. This
commit removes the pre-existing .md file in favor of a .html. It updates
the static asset handler in cored to serve the .html file as the index,
and also updates md2html to do the same.

Closes [unavailable]