core: use content-addressable snapshots
Split the `/rpc/get-snapshot` endpoint into two:

* `/rpc/get-snapshot-info` — returns height, size and blockchain ID for
latest snapshot
* `/rpc/get-snapshot` — retrieves a snapshot by its height

This helps with two things:

1. Once snapshots are content-addressable, we have the ability to cache
them like we're doing with get-blocks. We can do this later, post-launch if
needed.
2. The snapshot info endpoint will return the size of the snapshot in
bytes. We can use this information to show the user a progress bar in
dashboard as the snapshot downloads.

With this change, the `/rpc/get-snapshot` endpoint now returns the raw
protobuf instead of a base64-encoded string of the protobuf bytes.


Closes [unavailable]