core: generate a unique ID and provide it in RPCs
It would be nice to have some basic stats about how many Chain Cores are
using testnet. We now generate an ID when the core is configured and send
it along with each RPC. In the generator, we log the core ID if any.

Future work: collect basic stats such as how many distinct core IDs have
connected in the last five minutes.

Closes [unavailable]