core/migrate: work around pre-squash schemas
Some deployed systems are running a database from before
we squashed all the migrations, so we need to detect
that situation and handle it properly. If their schema
is equivalent to the current squash, we can replace
thier migration history with a single entry representing
the squash entry. This is the case if their last
migration is the last migration before the squash.

The squash was 836b05802c02b7d1cf991d7a29f668d26447cfe8
and the last migration was
2016-10-10.0.mockhsm.add-key-types.sql with hash
00ac8143767fe4a44855cab1ec57afd52c44fd4d727055db9e8584c3e9b10983.

Closes [unavailable]