dashboard: disable filter autocomplete on '='
This commit disables the `=''` helper in the filter bar.

As a programmer, I am used to having to type my own single quotes
after typing an equals sign. The current behavior might be too
helpful; my muscle memory is to type EQUALS and SINGLE QUOTE,
which causes the cursor to jump past the second single quote.

Closes [unavailable]