docs: fix doc index redirect by allowing HTML without layout
On Safari, font loading was broken by a JavaScript-powered redirect from
the index to the intro page. This can be solved by using an meta refresh
redirect.

To achieve the latter, md2html must support both HTML files that include
the layout (such as the install page), as well as HTML files that do
not include the layout (such as an index page with meta refresh). The
md2html has been updated such that source files ending with a
.partial.html extension are rendered with a layout, and all other HTML
files are rendered unmodified.

Closes [unavailable]