core/txbuilder: roll back builder side-effects
If one action succeeds with some side effects (most
notably, reserving some UTXOs) and another action fails,
we should roll back the side effects if possible so that
the overall operation has no effect.

Note that this is not a guaranteed operation, so any
side effects need to be designed to be safe not to roll
back. Reservations are: they will eventually expire on
their own. Meanwhile, it helps performance to eagerly
cancel them upon failure rather than waiting for the
deadline to pass.

Closes #94