docs/tools/checklinks.rb: fix reports for broken cross-file references
Lets say you have file `alice.md` with a cross-file link
`[link](bob.md#bad-anchor)`. 

Previous version of this tool was reporting:

    Broken link in bob.md: [link (from alice.md)](#bad-anchor)

Since usually it's the link that is broken, not the title, checklinks.rb
now reports:

    Broken link in alice.md: [link](bob.md#bad-anchor)

Which is much more useful: we know that we have to go to `alice.md` and fix
the link there.

Closes #129