core/pin: add block processors
This removes blocking callbacks from the protocol's commit block process
in favor of separate block processors that pull blocks. The processors
maintain pins that keep track of the last block that has been processed.
In the case of the cored process being restarted, the processors will
pick up where they last left off.

The advantage of separating block processors is that it allows the
validation and application of blocks to proceed unblocked by secondary
computations, such as indexing account utxos. It also paves the way for
some aspects of block processing to happen outside of the leader
process.

Closes #117