core: add release version to compile
This commit proposes a method for including version information into the
cored binary.

A new global constant, "version", is present in the cored executable.
It is returned by the /info endpoint.

When we have decided that the main branch contains the features and
bugfixes sufficient for a new release, we add a new commit that updates
the version constant. For posterity, we will also annotate this commit
with the git tag "cmd/cored-<VERSION>". When preparing new installers,
we can refer back to these tags to get a precise state of the
repository.

Initially, the version will be set to "?", but we should schedule an
uptick to 1.0.1 as soon as possible.

Worth noting: the version constant is useful for tagging public releases
and communicating with outside users. However, it is not very useful for
our internal development process, since the value of the version
constant will change infrequently. In these cases, other means of
communicating the source version (such as the buildcommit expvar, or the
local HEAD ref) should be used.