cmd/cored: add release version file
This gives us space to track the latest version of the cored binary in
version control.

When we have decided that the main branch contains the features and
bugfixes sufficient for a new release, we add a new commit that updates
the version in this file. Initially, the version will be set to "?", but
we should schedule an uptick to 1.0.1 as soon as possible.

For posterity, we will also annotate this commit with the git tag
"cmd/cored-<VERSION>". (If we want a nice, public UI to navigate the
release history, we might consider using GitHub's releases feature to
perform the tagging).

When compiling cored for a new release, use the contents of this file as
the buildTag build variable:

    buildTag=`cat cmd/cored/last-release-version | tr -d '[:space:]'`
    ldflags="-X main.buildTag=$buildTag ..."
    go build -ldflags "$ldflags" ...

The release version will only be included in builds that are built in
this manner.