cmd/md2html: multi-language code snippets
This commit adds multi-language snippet support to md2html. Documents
should provide snippet directives with the following structure:

    $code <snippet-name> <file1> <file2> ...

Example:

    $code create-client ../examples/java/FiveMinuteGuide.java
../examples/ruby/five_minute_guide.rb

A snippet block will be generated for every named file. The file must
contain the specified snippet, or a block of error text will be generated
instead.

In the browser, snippets for different languages will be displayed based on
the currently-selected language.

Closes #204