docs/internal: add Swagger version of API spec
This document is intended to provide comprehensive documentation
of Client API. Some of our partners have requested this style of
documentation. Since it is a superset of the information provided
in core/api-spec.md, it should eventually replace that document.

Unlike other Swagger API descriptions, this one is not intended to
be used for auto-generation of client libraries.

The current doc is incomplete; future PRs will fill out the rest of the
content.

Closes #230