cmd/md2html: fix HTML escaping in code snippets
This commit adds proper escaping of HTML entities in code snippets, a
bug that was introduced in when langage switching was added. This caused
issues such as incorrect rendering of Java generic expressions
("List<Asset>").

This bug was due to the following:

1. Adding switching behavior required adding logical divs that enclosed
the code snippets.
2. Enclosing divs cause blackfriday, the Markdown parser, to ignore
fenced code blocks inside those divs.
3. We switched to manually inserting <pre><code> wrappers, but neglected
to add HTML entity escaping.