cmd/md2html: fix HTML escaping in code snippets
This commit adds proper escaping of HTML entities in code snippets, a
bug that was introduced in when langage switching was added. This caused
issues such as incorrect rendering of Java generic expressions
(`List<Asset>`).

This bug was due to the following:

1. Adding switching behavior required adding logical divs that enclosed the
code snippets.
2. If we use enclosing divs, then fenced code blocks inside of those divs
will be ignored by the Markdown parser. This is part of the official
Markdown spec.
3. Without access to fenced code blocks, we switched to manually inserting
`<pre><code>` wrappers, but forgot to add HTML entity escaping.

Closes #327