protocol: remove Pool interface
The `protocol.Chain` no longer needs to know anything about the
pending transaction pool. We already only call `AddTx` on generator
nodes. We have an alternative method `ValidateTxCached` for just
performing the validation.

This removes the `protocol.Pool` interface and the `protocol.Chain.AddTx`
method. Instead, it introduces a `txbuilder.Submitter` type that is
implemented by `mempool.MemPool` and a wrapper around the remote
generator RPC client.

For now, `mempool.MemPool` still exists as a wedge between the
`core/txbuilder` package and the `core/generator`. In the future, we
may want to remove it and modify `core/generator.generator` to implement
`txbuilder.Submitter` directly. That will be helpful when modifying the
generator to build blocks incrementally as transactions are submitted to
the pool.

Note that this will also fix the behavior where if you setup your Core to
replicate with a non-generator Core, submitted txs would end up in the
non-generator Core's pool and not be propagated to the generator.

Closes #316