wercker.yml: disable flaky dashboard linting in CI
For reasons not yet fully understood, Wercker only intermittently
fails on lint errors on the dashboard source. The issue appears
connected to using a cached node_modules folder (this feature is
enabled by the inz/step-npm-install Wercker plugin). If this
cache is disabled, linter errors will appear consistently.

Caching node_modules is necessary for a smooth workflow, so we'll
disable automated linting for now, since it currently only serves
to intermittently fail other folks' PRs with irrelevent liting
errors.

As a fallback, we will make sure the team is diligent about self-linting
as they work. We're also planning to add a call to eslint from
bin/bundle-dashboard, so there is at least some form of
automated sanity-check.

Closes #363