protocol/mempool: delete package
Delete protocol/mempool and replace all of its uses with
core/generator.

For now, this relies on a new generator.Generator method PendingTxs that
returns the current pool. If we refactor the TestRecovery we might be
able to remove the PendingTxs method and add a MakeBlock method to
Generator to replace prottest.MakeBlock.

Closes #365