docker/ci: add protobuf to ci build
Protobuf 3.1 is not added until alpine 3.5, so that repository has to
be added to the list to get it. Additionally, protobuf was compiled
using glibc, but alpine by default only has musl. A third party
maintains an alpine package providing musl.

A patch that allows protobuf to be compiled with musl has already been
merged, and should be available at the next release of protobuf. At that
time, glibc should be removed from this build.