all: don't require GOPATH to be set
Starting in Go 1.8, it is possible to install and use Go
without setting GOPATH. Several of our shell scripts and
instructions assume that GOPATH is set (almost always to
determine the location of installed executables), but
this assumption will become invalid soon. (And I think
it's unreasonable to impose setting GOPATH as an extra
requirement just for our environment.) We can fix this
by calling 'go env GOPATH', which prints the computed
value of GOPATH (either the string the user set, or the
default, which happens to be $HOME/go in Go 1.8).

Closes #401