testutil: add func DeepEqual, treats empty as equal to nil
The new function treats `nil` as equal to zero-length slices and maps, and
is otherwise mostly the same as `reflect.DeepEqual`. This slightly more
relaxed definition of equality should reduce the time wasted on
hard-to-track-down encoding errors that sometimes convert e.g. `[]byte{}`
to `nil` and vice versa.

This PR also replaces uses of `reflect.DeepEqual` in `core` and `protocol`
with `testutil.DeepEqual`.

Closes #420