docs: change versioning rules
To simplify the release of new features and updates,, we are
relaxing the compatibility constraints so that only software
packages with identitcal major and minor versions are considered
mutually-compatible. This commit removes the "tolerance" language
from the versioning doc; compatibility is derived from strict
equality of major and minor versions.

This allows us, among other things, to introduce breaking changes
without undergoing a deprecation cycle. This not only includes
user-facing breaking changes, but also changes in the wire
protocol. For example, if the wire representation of a particular
API call changes from A to A', then the tolerance rule would force
us accept A or A' for a whole minor version.

This also resolves a subtle ambiguity in mutual compatability
between the client SDK and the server API. Even when applying a
standard deprecation process, we cannot guarantee both of the
following:

1) SDK 1.0 is compatible with server 1.1
2) SDK 1.1 is compatible with server 1.0

Using the wire protocol example above, even if server 1.1 supports
both the old (A) and new (A') API call, SDK 1.1 will have to
adopt A'. This will break compatibility between SDK 1.1 and server
1.0.