core/query/filter: add unicode support
Support UTF-8 characters in strings and identifiers. This does _not_ add
escaping for strings or identifiers so `'` and `\` are still illegal
characters in strings. Any non-alphanumeric characters are still
illegal in identifiers.

Token positions are reported using the byte offset of the beginning of
the token. To retrieve the code point offset, the consumer of the package
can use `utf8.RuneCount(src[:offset])`.

Closes #445